Adds oath_server_client_info to apply_auth
