+++ modules/openapi-json-schema-generator/src/main/resources/python/security_schemes.hbs
@@ -14,6 +14,26 @@ from urllib3 import _collections
 
 from {{packageName}} import signing
 {{/if}}
+{{#if oauthServerHostnames}}
+
+
+@dataclasses.dataclass
+class OauthClientInfo:
+    client_id: str
+    client_secret: typing.Optional[str] = None
+
+
+# oauth server to client info
+OauthServerClientInfo = typing_extensions.TypedDict(
+    'OauthServerClientInfo',
+    {
+{{#each oauthServerHostnames}}
+        "{{{.}}}": OauthClientInfo,
+{{/each}}
+    },
+    total=False
+)
+{{/if}}
 
 
 class SecuritySchemeType(enum.Enum):
@@ -39,6 +59,9 @@ class __SecuritySchemeBase(metaclass=abc.ABCMeta):
         method: str,
         body: typing.Optional[typing.Union[str, bytes]],
         scope_names: typing.Tuple[str] = (),
+{{#if oauthServerHostnames}}
+        oath_server_client_info: OauthServerClientInfo = {}
+{{/if}}
     ) -> None:
         pass
 
@@ -57,6 +80,9 @@ class ApiKeySecurityScheme(__SecuritySchemeBase, abc.ABC):
         method: str,
         body: typing.Optional[typing.Union[str, bytes]],
         scope_names: typing.Tuple[str] = (),
+{{#if oauthServerHostnames}}
+        oath_server_client_info: OauthServerClientInfo = {}
+{{/if}}
     ) -> None:
         if self.in_location is ApiKeyInLocation.COOKIE:
             headers.add('Cookie', self.api_key)
@@ -93,6 +119,9 @@ class HTTPBasicSecurityScheme(__SecuritySchemeBase):
         method: str,
         body: typing.Optional[typing.Union[str, bytes]],
         scope_names: typing.Tuple[str] = (),
+{{#if oauthServerHostnames}}
+        oath_server_client_info: OauthServerClientInfo = {}
+{{/if}}
     ) -> None:
         user_pass = f"{self.user_id}:{self.password}"
         b64_user_pass = base64.b64encode(user_pass.encode(encoding=self.encoding))
@@ -113,6 +142,9 @@ class HTTPBearerSecurityScheme(__SecuritySchemeBase):
         method: str,
         body: typing.Optional[typing.Union[str, bytes]],
         scope_names: typing.Tuple[str] = (),
+{{#if oauthServerHostnames}}
+        oath_server_client_info: OauthServerClientInfo = {}
+{{/if}}
     ) -> None:
         headers.add('Authorization', f"Bearer {self.access_token}")
 {{#if hasHttpSignatureSecurityScheme}}
@@ -131,6 +163,9 @@ class HTTPSignatureSecurityScheme(__SecuritySchemeBase):
         method: str,
         body: typing.Optional[typing.Union[str, bytes]],
         scope_names: typing.Tuple[str] = (),
+{{#if oauthServerHostnames}}
+        oath_server_client_info: OauthServerClientInfo = {}
+{{/if}}
     ) -> None:
         query_params = tuple()
         auth_headers = self.signing_info.get_http_signature_headers(
@@ -152,6 +187,9 @@ class HTTPDigestSecurityScheme(__SecuritySchemeBase):
         method: str,
         body: typing.Optional[typing.Union[str, bytes]],
         scope_names: typing.Tuple[str] = (),
+{{#if oauthServerHostnames}}
+        oath_server_client_info: OauthServerClientInfo = {}
+{{/if}}
     ) -> None:
         raise NotImplementedError("HTTPDigestSecurityScheme not yet implemented")
 
@@ -167,6 +205,9 @@ class MutualTLSSecurityScheme(__SecuritySchemeBase):
         method: str,
         body: typing.Optional[typing.Union[str, bytes]],
         scope_names: typing.Tuple[str] = (),
+{{#if oauthServerHostnames}}
+        oath_server_client_info: OauthServerClientInfo = {}
+{{/if}}
     ) -> None:
         raise NotImplementedError("MutualTLSSecurityScheme not yet implemented")
 
@@ -222,6 +263,9 @@ class OAuth2SecurityScheme(__SecuritySchemeBase, abc.ABC):
         method: str,
         body: typing.Optional[typing.Union[str, bytes]],
         scope_names: typing.Tuple[str] = (),
+{{#if oauthServerHostnames}}
+        oath_server_client_info: OauthServerClientInfo = {}
+{{/if}}
     ) -> None:
         raise NotImplementedError("OAuth2SecurityScheme not yet implemented")
 
@@ -237,6 +281,9 @@ class OpenIdConnectSecurityScheme(__SecuritySchemeBase, abc.ABC):
         method: str,
         body: typing.Optional[typing.Union[str, bytes]],
         scope_names: typing.Tuple[str] = (),
+{{#if oauthServerHostnames}}
+        oath_server_client_info: OauthServerClientInfo = {}
+{{/if}}
     ) -> None:
         raise NotImplementedError("OpenIdConnectSecurityScheme not yet implemented")
